+++ src/main/java/dev/muon/otherworldorigins/ModEvents.java
@@ -22,4 +22,5 @@ public class ModEvents {
     public static void modifyEntityAttributes(EntityAttributeModificationEvent event) {
         event.add(EntityType.PLAYER, ModAttributes.HEALTH_PER_LEVEL.get());
     }
+
 }